bake-decouple: move bake source-provider host exports into bake
